+++ src/util.h
@@ -12,7 +12,7 @@ void particles_gen(double **P, int light_cnt, int medium_cnt, int heavy_cnt, int
 // Update particle
 void update(unsigned char* image, double **P, double **P_force,
 	int total_p_cnt, int img_width, int img_height,
-	double step_size, int regenerate_img);
+	int step_size, int regenerate_img);
